presets: move preset and preset-io into panel-plugin/presets/
