+++ panel-plugin/settings-dialog.cpp
@@ -27,8 +27,8 @@
 #include "settings.h"
 #include "slot.h"
 #include "window.h"
-#include "preset.h"
-#include "preset-io.h"
+#include "presets/preset.h"
+#include "presets/preset-io.h"
 #include "unified-bar.h"
 #include "ui/properties/common.h"
